requirements.txt
tensorflow==2.3.0
Keras==2.4.3
matplotlib==3.3.4
numpy==1.19.2
opencv_python==4.5.1.48
pandas==1.2.1
Pillow==8.1.0
scikit_image==0.18.1
XlsxWriter==1.3.7
protobuf<3.20
tifffile==2021.1.14
